requirements.txt
allure-python-commons
allure-pytest
pytest
pytest-html
requests
selenium